move waitForTcpPortOpen to port_utils
